+++ spec/rubocop/cop/cop_spec.rb
@@ -95,7 +95,7 @@ describe RuboCop::Cop::Cop do
       context 'when offense was corrected' do
         before do
           allow(@cop).to receive(:autocorrect?).and_return(true)
-          allow(@cop).to receive(:autocorrect).and_return(->(corrector) {})
+          allow(@cop).to receive(:autocorrect).and_return(->(_corrector) {})
         end
 
         it 'is set to true' do